+++ test/unit/readme.test.ts
@@ -26,6 +26,33 @@ describe('readme', () => {
     expect(fs.readFileSync('README.md', 'utf8')).to.contain('multi')
   })
 
+  describe('with command that has an alias', () => {
+    const rootPath = path.join(__dirname, '../fixtures/cli-command-with-alias')
+    const readmePath = path.join(rootPath, 'README.md')
+    const originalReadme = fs.readFileSync(readmePath, 'utf8')
+    const aliasOutput = '`oclif hi`'
+
+    test
+    .stdout()
+    .finally(() => fs.writeFileSync(readmePath, originalReadme))
+    .stub(process, 'cwd', () => rootPath)
+    .command(['readme'])
+    .it('--aliases flag (default)', () => {
+      const newReadme = fs.readFileSync(readmePath, 'utf8')
+      expect(newReadme).to.contain(aliasOutput)
+    })
+
+    test
+    .stdout()
+    .finally(() => fs.writeFileSync(readmePath, originalReadme))
+    .stub(process, 'cwd', () => rootPath)
+    .command(['readme', '--no-aliases'])
+    .it('--no-aliases flag', () => {
+      const newReadme = fs.readFileSync(readmePath, 'utf8')
+      expect(newReadme).not.to.contain(aliasOutput)
+    })
+  })
+
   describe('with custom help that implements formatCommand', () => {
     const rootPath = path.join(__dirname, '../fixtures/cli-with-custom-help')
     const readmePath = path.join(rootPath, 'README.md')