+++ CHANGELOG.md
@@ -2,6 +2,93 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [3.2.2](https://github.com/oclif/oclif/compare/v3.2.1...v3.2.2) (2022-09-16)
+
+### [3.2.1](https://github.com/oclif/oclif/compare/v3.2.0...v3.2.1) (2022-09-09)
+
+
+### Bug Fixes
+
+* bump oclif/core ([21aa0d7](https://github.com/oclif/oclif/commit/21aa0d754f0afe5a2e9f25cd1c32266ef2b9ac2d))
+
+## [3.2.0](https://github.com/oclif/oclif/compare/v3.1.2...v3.2.0) (2022-09-06)
+
+
+### Features
+
+* add `--aliases` flag for readme command ([#925](https://github.com/oclif/oclif/issues/925)) ([#952](https://github.com/oclif/oclif/issues/952)) ([5bf2b5a](https://github.com/oclif/oclif/commit/5bf2b5a51210d554a1768d69fbf93c9c6f6b2b2a))
+
+### [3.1.2](https://github.com/oclif/oclif/compare/v3.1.1...v3.1.2) (2022-07-27)
+
+
+### Bug Fixes
+
+* await yeoman generators ([#926](https://github.com/oclif/oclif/issues/926)) ([da5a4d2](https://github.com/oclif/oclif/commit/da5a4d202be114fe7f0c00d2737a9faf8f4e19ab))
+
+### [3.1.1](https://github.com/oclif/oclif/compare/v3.1.0...v3.1.1) (2022-07-15)
+
+
+### Bug Fixes
+
+* readme for single command cli ([#916](https://github.com/oclif/oclif/issues/916)) ([242b34f](https://github.com/oclif/oclif/commit/242b34f62656b2c6ee8a1d0ecda8d98f8c1c5bbc))
+
+## [3.1.0](https://github.com/oclif/oclif/compare/v3.0.1...v3.1.0) (2022-07-07)
+
+
+### Features
+
+* build/pack in parallel ([#914](https://github.com/oclif/oclif/issues/914)) ([d86c722](https://github.com/oclif/oclif/commit/d86c722d8de66560c672c2fbdb338cc9a024caa2))
+
+### [3.0.1](https://github.com/oclif/oclif/compare/v3.0.0...v3.0.1) (2022-04-18)
+
+
+### Bug Fixes
+
+* remove outdated code ([#866](https://github.com/oclif/oclif/issues/866)) ([e3f60a3](https://github.com/oclif/oclif/commit/e3f60a3109502bc59d1fbd71a3fbbd3668dcd3eb))
+
+## [3.0.0](https://github.com/oclif/oclif/compare/v2.6.3...v3.0.0) (2022-04-13)
+
+
+### ⚠ BREAKING CHANGES
+
+* Add arch to macos pkg file (#849)
+
+### Features
+
+* Add arch to macos pkg file ([#849](https://github.com/oclif/oclif/issues/849)) ([656d529](https://github.com/oclif/oclif/commit/656d52909c0ae1c4b66b7b3eaada9916d5871e4a))
+
+### [2.6.3](https://github.com/oclif/oclif/compare/v2.6.2...v2.6.3) (2022-03-31)
+
+
+### Bug Fixes
+
+* update oclif/core ([7e86036](https://github.com/oclif/oclif/commit/7e86036b4e673003ad6226852260c929d62a1c39))
+
+### [2.6.2](https://github.com/oclif/oclif/compare/v2.6.1...v2.6.2) (2022-03-31)
+
+
+### Bug Fixes
+
+* revert building mac m1 packages, wait for PR [#849](https://github.com/oclif/oclif/issues/849) ([#855](https://github.com/oclif/oclif/issues/855)) ([e5b68af](https://github.com/oclif/oclif/commit/e5b68af8510b0419ec48d76087ed0bec61fc81ad))
+
+### [2.6.1](https://github.com/oclif/oclif/compare/v2.6.0...v2.6.1) (2022-03-23)
+
+## [2.6.0](https://github.com/oclif/oclif/compare/v2.5.0...v2.6.0) (2022-03-14)
+
+
+### Features
+
+* Add support to build darwin-arm64 for `oclif pack macos` ([#847](https://github.com/oclif/oclif/issues/847)) ([61a96c2](https://github.com/oclif/oclif/commit/61a96c2a1c435f5abce8c2348cf7756a4b997269))
+
+## [2.5.0](https://github.com/oclif/oclif/compare/v2.4.5...v2.5.0) (2022-02-24)
+
+
+### Features
+
+* remove all relevant directories on uninstall ([#831](https://github.com/oclif/oclif/issues/831)) ([acd986d](https://github.com/oclif/oclif/commit/acd986d53681b3840f17220b6e16fd8bb3c869f7))
+
+### [2.4.5](https://github.com/oclif/oclif/compare/v2.4.4...v2.4.5) (2022-02-17)
+
 ### [2.4.4](https://github.com/oclif/oclif/compare/v2.4.3...v2.4.4) (2022-02-09)